Moving location of zuul api interfaces.
